third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/max-length-for-device-name-namePrefix.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let name_too_long = generate_string(249, 'a');
  return assert_promise_rejects_with_message(
    requestDeviceWithKeyDown({filters: [{namePrefix: name_too_long}]}),
      new DOMException(
        'Failed to execute \'requestDevice\' on \'Bluetooth\': A device ' +
        'name can\'t be longer than 248 bytes.',
      'TypeError'),
    'Device name longer than 29');
}, 'A device name prefix longer than 248 must reject.');
</script>
